refactor(plugin-inbox,pa): port the real inbox domain into plugin-inbox — aggregate, priority scoring, routes, triage (#8652's biggest code item)
#8652's reopen finding: plugin-inbox was supposed to carry the real inbox
operations but #8670 only registered a context provider. The domain now
actually moves: aggregate (977 lines) + priority-scoring (442) + message
fetching + types live in plugins/plugin-inbox with correct dependency
direction (zero PA imports); the PA files become thin re-export shims so
every existing consumer keeps working (importers verified). LifeOps
commandments respected: no second store, typed DispatchResult, structural
routing.

Real-runtime route + aggregate suites (433+481 lines) drive the real
service against a real runtime harness; PA inbox-triage integration suite
stays green. Adversarially verified (approve — 'genuine domain move, fully
substantiated').
