+++ .github/issue-evidence/8652-plugin-inbox-port/live-cerebras-priority-scoring.json
@@ -1,0 +1,70 @@
+{
+  "model": "gpt-oss-120b",
+  "messages": [
+    {
+      "id": "m1",
+      "channel": "gmail",
+      "from": "Acme Billing",
+      "snippet": "Invoice #4482 is 10 days overdue — service suspension on Friday unless paid."
+    },
+    {
+      "id": "m2",
+      "channel": "discord",
+      "from": "gamer_pal",
+      "snippet": "yo did you see the new patch notes lol"
+    },
+    {
+      "id": "m3",
+      "channel": "gmail",
+      "from": "Dr. Chen's office",
+      "snippet": "Reminder: your appointment is Tuesday at 9am. Reply C to confirm."
+    },
+    {
+      "id": "m4",
+      "channel": "telegram",
+      "from": "Mom",
+      "snippet": "Call me when you can, nothing urgent, just want to plan Sunday dinner."
+    }
+  ],
+  "scores": [
+    {
+      "score": 90,
+      "category": "important",
+      "flags": [
+        "money",
+        "deadline",
+        "urgent"
+      ]
+    },
+    {
+      "score": 20,
+      "category": "casual",
+      "flags": []
+    },
+    {
+      "score": 78,
+      "category": "planning",
+      "flags": [
+        "meeting",
+        "question"
+      ]
+    },
+    {
+      "score": 70,
+      "category": "planning",
+      "flags": [
+        "ask",
+        "mention"
+      ]
+    }
+  ],
+  "trajectory": {
+    "model": "gpt-oss-120b",
+    "calls": [
+      {
+        "prompt": "Score each inbox message. For each one decide:\n- score: integer 0–100 reflecting how much the user should care right now (100 = drop everything; 0 = ignorable).\n- category: one of:\n    - important: high signal, demands attention or action soon\n    - planning: schedules, dates, times, RSVPs, meeting coordination\n    - casual: chit-chat, social, low-stakes notifications\n- flags: zero or more short tags from { mention, question, deadline, meeting, money, urgent, group_call, ask }\n\nCalibration:\n- Direct DMs from real people > group chatter > newsletters/automation.\n- Mentions of the user, direct questions, or scheduling/dates push score above 70.\n- Pure social pleasantries with no ask and no time element belong in 'casual' with score < 35.\n\nOwner context:\nownerName: Shaw\n\nMessages:\nmessages[0]:\n  from: Acme Billing\n  channel: gmail\n  chatType: dm\n  participantCount: null\n  subject: OVERDUE: Invoice #4482\n  snippet: Invoice #4482 is 10 days overdue — service suspension on Friday unless paid.\nmessages[1]:\n  from: gamer_pal\n  channel: discord\n  chatType: dm\n  participantCount: null\n  subject: null\n  snippet: yo did you see the new patch notes lol\nmessages[2]:\n  from: Dr. Chen's office\n  channel: gmail\n  chatType: dm\n  participantCount: null\n  subject: Appointment reminder\n  snippet: Reminder: your appointment is Tuesday at 9am. Reply C to confirm.\nmessages[3]:\n  from: Mom\n  channel: telegram\n  chatType: dm\n  participantCount: null\n  subject: null\n  snippet: Call me when you can, nothing urgent, just want to plan Sunday dinner.\n\nReturn JSON only as a single object. No prose, markdown, code fences, or hidden reasoning.\nReturn exactly 4 zero-based scores records aligned to messages[0] through messages[3].\nUse this shape: {\"scores\":[{\"score\":82,\"category\":\"important\",\"flags\":[\"question\",\"deadline\"]}]}\nscore is an integer 0-100. category is important, planning, or casual.\nflags is an array of tags from mention, question, deadline, meeting, money, urgent, group_call, ask.",
+        "rawResponse": "{\"scores\":[{\"score\":90,\"category\":\"important\",\"flags\":[\"money\",\"deadline\",\"urgent\"]},{\"score\":20,\"category\":\"casual\",\"flags\":[]},{\"score\":78,\"category\":\"planning\",\"flags\":[\"meeting\",\"question\"]},{\"score\":70,\"category\":\"planning\",\"flags\":[\"ask\",\"mention\"]}]}"
+      }
+    ]
+  }
+}
