+++ .github/issue-evidence/8652-plugin-inbox-port/registration-dedup-proof.txt
@@ -1,0 +1,51 @@
+# Proof: PA's inbox action/provider registrations were dead duplicates at runtime
+
+## 1. Core registration order (packages/core/src/runtime.ts, develop 5471346e7a6)
+
+registerPlugin() executes strictly in this order (excerpted):
+
+  ~line 1880:  if (pluginToRegister.init) { ... await pluginToRegister.init(config, this); ... }
+  ~line 1902:  if (pluginToRegister.actions) { ... existingActionNames.has(action.name) ->
+               "Skipping duplicate plugin action" ... }
+  ~line 1923:  if (pluginToRegister.providers) { ... existingProviderNames.has(provider.name) ->
+               "Skipping duplicate plugin provider" ... }
+
+i.e. a plugin's `init()` runs BEFORE its own `actions`/`providers` arrays register.
+
+## 2. PA init loads plugin-inbox first
+
+plugins/plugin-personal-assistant/src/plugin.ts init():
+  await ensureLifeOpsInboxPluginRegistered(runtime);   // registers the whole inboxPlugin
+That call happens inside PA's init(), so by the time the runtime walks PA's own
+`actions`/`providers` arrays, plugin-inbox has ALREADY registered:
+  - the INBOX umbrella + INBOX_* promoted virtuals (post-port, promoted by
+    plugin-inbox itself in plugins/plugin-inbox/src/plugin.ts), and
+  - the `inboxTriage` provider.
+
+Consequence (pre-change develop): PA's own `inboxTriageProvider` array entry and
+`...promoteSubactionsToActions(inboxAction)` spread hit the duplicate-name guard
+("Skipping duplicate plugin provider" / "Skipping duplicate plugin action") on
+every boot — dead code. PA's local 179-line provider copy in
+src/providers/inbox-triage.ts was never executed.
+
+## 3. Post-change: single-source registration, observable in a real runtime
+
+plugins/plugin-inbox/test/inbox-routes.real-runtime.test.ts registers the REAL
+inboxPlugin STANDALONE (no PA) on a real PGLite AgentRuntime and asserts:
+
+  - runtime.actions contains INBOX, INBOX_LIST, INBOX_TRIAGE, INBOX_REPLY,
+    INBOX_SNOOZE, INBOX_ARCHIVE, INBOX_APPROVE
+  - runtime.providers contains inboxTriage + inboxCrossChannelContext
+
+(pre-port, standalone plugin-inbox registered only the raw INBOX umbrella; the
+INBOX_* virtuals existed only when PA was loaded — the standalone gap this port
+closes). See test-plugin-inbox.txt for the green run.
+
+## 4. Egress-redaction equivalence of the removed PA provider copy
+
+PA's dead copy wrapped snippets/deep-links in LifeOps egress redaction. It built
+its context as createLifeOpsEgressContext({ isOwner: true, ... }) and the
+provider is owner-gated twice (roleGate.minRole OWNER + hasOwnerAccess()), so
+every reachable call site redacted with an owner context — a pass-through. The
+live plugin-inbox provider (also roleGate OWNER + hasOwnerAccess) is observably
+identical for every reachable input.
