+++ .github/issue-evidence/8652-plugin-inbox-port/test-pa-inbox-subset.txt
@@ -1,0 +1,19 @@
+
+ RUN  v4.1.5 /Users/shawwalters/eliza-workspace/milady/eliza/.claude/worktrees/ui-mobile-wave3
+
+
+ Test Files  3 passed (3)
+      Tests  19 passed (19)
+   Start at  04:13:46
+   Duration  12.90s (transform 8.27s, setup 195ms, import 12.42s, tests 33ms, environment 0ms)
+
+--- integration lane: lifeops-inbox-triage ---
+
+ RUN  v4.1.5 /Users/shawwalters/eliza-workspace/milady/eliza/.claude/worktrees/ui-mobile-wave3
+
+
+ Test Files  1 passed (1)
+      Tests  4 passed (4)
+   Start at  04:13:59
+   Duration  46.19s (transform 27.46s, setup 52ms, import 9.02s, tests 37.01s, environment 0ms)
+
