+++ .github/issue-evidence/8652-plugin-inbox-port/typecheck.txt
@@ -1,0 +1,7 @@
+== plugin-inbox: tsgo --noEmit -p tsconfig.json ==
+$ tsgo --noEmit -p tsconfig.json
+exit: 0
+
+== plugin-personal-assistant: tsc --noEmit -p tsconfig.build.json ==
+$ tsc --noEmit -p tsconfig.build.json
+exit: 0
