+++ plugins/plugin-inbox/src/plugin.ts
@@ -1,4 +1,4 @@
-import type { Plugin } from "@elizaos/core";
+import { type Plugin, promoteSubactionsToActions } from "@elizaos/core";
 
 import { inboxAction } from "./actions/inbox.ts";
 import { inboxDbSchema } from "./db/schema.ts";
@@ -10,11 +10,13 @@ import { inboxRoutes } from "./routes/inbox-routes.ts";
 export const inboxPlugin: Plugin = {
   name: "@elizaos/plugin-inbox",
   description:
-    "Unified cross-channel inbox triage with unresolved-item tracking. Hosts the INBOX umbrella action (list/search/summarize fan-out across email/Discord/Telegram/WhatsApp/X/Slack and similar non-SMS channels) and the inboxTriage provider, backed by the InboxService/InboxRepository triage back-end. The cross-channel inbox read route (`GET /api/lifeops/inbox`) and the connector-coupled getInbox/cross-channel-context surfaces stay in @elizaos/plugin-personal-assistant, which delegates the triage domain here. (Android SMS is handled by plugin-messages.)",
+    "Unified cross-channel inbox triage with unresolved-item tracking. Hosts the INBOX umbrella action (list/search/summarize fan-out across email/Discord/Telegram/WhatsApp/X/Slack and similar non-SMS channels) and the inboxTriage provider, backed by the InboxService/InboxRepository triage back-end plus the aggregation domain in `inbox/aggregate.ts` (builders, request resolver, cached read-through InboxDomain). The legacy transport route `GET /api/lifeops/inbox` and the connector sources/cache tables stay in @elizaos/plugin-personal-assistant, which injects them through the aggregate seams and delegates the domain here. (Android SMS is handled by plugin-messages.)",
   dependencies: ["@elizaos/plugin-sql"],
   schema: inboxDbSchema,
   services: [InboxMigrationService],
-  actions: [inboxAction],
+  // Promote the INBOX_* subaction virtuals here so they exist wherever the
+  // plugin loads (including standalone, without plugin-personal-assistant).
+  actions: [...promoteSubactionsToActions(inboxAction)],
   providers: [inboxTriageProvider, crossChannelContextProvider],
   routes: inboxRoutes,
   views: [
